+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Version [0.9.0](https://github.com/cedx/base/compare/v0.8.0...v0.9.0)
 - Added the `Hide()` and `Show()` methods to the `LoadingIndicator` and `OfflineIndicator` UI components.
 - Added the `UI.StorageArea` enumeration.
+- Added the `UI.Components.TabActivator` component.
 
 ## Version [0.8.0](https://github.com/cedx/base/compare/v0.7.0...v0.8.0)
 - Added the `Hosting.Environment` enumeration.